+++ .gitignore
@@ -1,2 +1,7 @@
 .claude/
 .obsidian/
+node_modules/
+test-results/
+playwright-report/
+playwright/.cache/
+*.log
