+++ README.md
@@ -1,20 +1,18 @@
 # Reddit Translator Pro Auto
 
-`Reddit Translator Pro Auto` 是一个面向 Reddit 页面使用的用户脚本，适用于 `Tampermonkey` 和 `Violentmonkey`。它用于翻译帖子正文、评论内容和滚动浏览中的可见文本，并提供多引擎、双语阅读、历史记录、快捷键和自动翻译控制能力。
+`Reddit Translator Pro Auto` 是一个面向 `Reddit` 页面使用的用户脚本，适用于 `Tampermonkey` 和 `Violentmonkey`。它用于翻译帖子正文、评论内容和滚动浏览中的可见文本，并提供多引擎、双语阅读、历史记录、快捷键、缓存和自动翻译控制能力。
 
-当前仓库的正式发布文件是：
+## 当前状态
 
-- `reddit-translator-pro-auto.user.js`
+- 当前开发源：`Reddit 翻译器 Pro 修6.js`
+- 当前正式发布文件：`reddit-translator-pro-auto.user.js`
+- 当前正式发布版本：`v1.0.8`
 
-当前正式发布版本是：
+仓库中的开发、发布和历史文件现在按职责分开维护：
 
-- `v1.0.7`
-
-当前发布文件已经与：
-
-- `Reddit 翻译器 Pro 修5.js`
-
-完全同步。
+- `Reddit 翻译器 Pro 修6.js` 用于继续开发、调试和功能迭代
+- `reddit-translator-pro-auto.user.js` 用于安装、分发和对外发布
+- `archive/` 用于保留旧阶段快照，不作为当前安装入口
 
 ## 功能概览
 
@@ -22,116 +20,89 @@
 - 支持 100+ 目标语言
 - 支持帖子正文与评论内容翻译
 - 支持双语显示与原文 / 译文切换
-- 支持滚动时自动翻译
-- 支持长文本和多段文本自动分块翻译
+- 支持滚动场景中的自动翻译
+- 支持长文本与多段内容自动分块翻译
 - 支持请求并发、速率、单次文本长度、单次段落数限制
-- 支持 TTS、历史记录、缓存、隐身模式
-- 支持主题切换、自定义颜色与快捷键
-- 支持 DeepL Key 保存、测试与轮换
+- 支持 `TTS`、历史记录、缓存、隐身模式
+- 支持主题切换、自定义颜色和快捷键
+- 支持 `DeepL API Key` 保存、测试和轮换
 
-## 当前版本重点
+## 仓库结构
 
-`v1.0.7` 以修5为正式发布基线，核心价值不是新增一批界面按钮，而是把修4已经完成的关键能力稳定下来，并整理为可持续发布的 `auto.user` 版本。
-
-这一版本的实际可用能力包括：
-
-- 请求限制控制：可配置最大并发请求数、每秒最大请求数、每次请求最大文本长度、每次请求最大段落数
-- 分块翻译：长文本和多段内容会按限制自动拆分再翻译，降低失败率
-- 视图切换：已翻译内容支持快速在原文与译文之间切换
-- 自动滚动翻译：在滚动浏览 Reddit 内容时，脚本可对进入视口的内容自动翻译
-- 历史与缓存：保留最近翻译记录，并通过缓存减少重复请求
-- 持久化隔离：修5使用新的存储前缀，减少修4阶段旧配置对当前版本的影响
-
-## 版本演进
-
-### 修1
-
-- 建立基础脚本结构
-- 支持 `Google` 和 `MyMemory`
-- 提供主题、双语、TTS、历史、快捷键等基础功能
-
-### 修2
-
-- 新增 `DeepL` 支持
-- 新增 DeepL API Key 配置能力
-- 将脚本从双引擎扩展为三引擎
-
-### 修3
-
-- 整合前两版能力
-- 调整默认配置
-- 为后续请求调度和自动翻译增强做结构准备
-
-### 修4
-
-- 新增请求限制设置
-- 新增长文本 / 多段文本自动分块翻译
-- 新增原文 / 译文切换
-- 强化自动滚动翻译触发逻辑
-- 将请求流程升级为并发与速率可控模型
-
-### 修5
-
-- 版本提升到 `v1.0.7`
-- 整理正式发布信息
-- 使用新的脚本存储前缀隔离旧版配置
-- 与 `reddit-translator-pro-auto.user.js` 对齐，作为当前发布基线
-
-## 文件说明
-
-### 正式发布文件
-
+- `Reddit 翻译器 Pro 修6.js`
+  当前开发源文件。后续功能修改优先落在这里。
 - `reddit-translator-pro-auto.user.js`
-
-这是安装、分发和后续更新说明应使用的唯一主文件。
-
-### 阶段开发文件
-
-- `Reddit 翻译器 Pro 修1.js`
-- `Reddit 翻译器 Pro 修2.js`
-- `Reddit 翻译器 Pro 修3.js`
-- `Reddit 翻译器 Pro 修4.js`
-- `Reddit 翻译器 Pro 修5.js`
-
-这些文件用于保留功能演进过程和阶段版本，不建议作为最终安装入口。
-
-### 早期历史文件
-
-- `reddit-translator-v1_0_0.js`
-
-该文件用于保留仓库早期历史版本，便于对照，不作为当前正式发布版本。
+  当前正式发布产物。用于安装、分发、GreasyFork 更新和版本说明。
+- `archive/`
+  存放 `修1` 到 `修5` 以及更早的历史快照，仅用于回溯和对照。
+- `tests/`
+  Playwright 冒烟测试。
+- `scripts/`
+  本地浏览器启动脚本。
+- `更新日志.md`
+  发布记录与阶段演进说明。
 
 ## 安装方式
 
-1. 在浏览器中安装用户脚本扩展，例如 `Tampermonkey` 或 `Violentmonkey`
+1. 在浏览器中安装 `Tampermonkey` 或 `Violentmonkey`
 2. 安装或导入 `reddit-translator-pro-auto.user.js`
-3. 打开 `https://www.reddit.com/` 页面
-4. 使用脚本面板设置目标语言、翻译引擎、请求限制和自动翻译行为
+3. 打开 `https://www.reddit.com/`
+4. 按 `F2` 打开脚本面板并完成语言、引擎与自动翻译设置
 
 ## 使用方法
 
-### 在脚本设置页面开启功能
+1. 进入任意 Reddit 帖子页、评论页或首页信息流
+2. 点击帖子或评论旁的翻译按钮翻译当前内容
+3. 在设置面板中选择目标语言、翻译引擎和双语模式
+4. 如需长文本稳定性，可调整请求并发、请求频率、单次字符数和单次段落数
+5. 开启自动翻译后，脚本会在内容进入视口时自动尝试翻译
+6. 翻译完成后，可在原文和译文之间切换
 
-1. 打开任意 `Reddit` 页面后，按 `F2` 打开脚本设置面板。
-2. 如果当前自定义过快捷键，也可以使用你在面板里设置的对应按键打开面板。
-3. 在设置页中选择目标语言和翻译引擎。
-4. 如果使用 `DeepL`，先填写 API Key，并在面板中测试可用性。
-5. 如需多个 `DeepL` Key，可使用 `,` 分隔填写。
-6. 根据需要开启双语显示、滚动自动翻译、隐身模式、TTS 等功能开关。
-7. 如果翻译长帖或长评论，可在请求限制区域设置并发数、请求频率、单次文本长度和单次段落数。
-8. 设置完成后关闭面板，脚本会按当前配置在页面中运行。
+## 当前发布线说明
 
-### 页面中的使用方式
+`v1.0.8` 是当前正式发布版本。发布文件由开发源整理后产出，当前发布线的重点包括：
 
-1. 打开任意 `Reddit` 帖子页、评论页或首页信息流。
-2. 点击帖子或评论旁的翻译按钮，翻译当前内容。
-3. 需要批量翻译时，可使用“翻译全部”按钮，或按默认快捷键 `Ctrl+Shift+T`。
-4. 开启滚动自动翻译后，脚本会在内容进入视口时自动尝试翻译。
-5. 翻译完成后，可通过原文 / 译文切换按钮在不同视图之间切换。
-6. 如果开启双语模式，页面会同时保留原文和译文，便于对照阅读。
+- 发布入口统一为 `reddit-translator-pro-auto.user.js`
+- 默认体验偏向开箱即用，发布文件内的默认配置已针对自动翻译场景做过整理
+- 请求调度、分块翻译、滚动触发、缓存和历史记录能力已进入稳定可用阶段
+- 发布线与历史快照分离，减少仓库根目录中的版本噪音
 
-## 更新与维护原则
+开发源与发布产物不是“永远字节级完全相同”的双份主文件，而是“一个负责开发，一个负责发布”的关系。后续文档、安装说明和外部链接都应以 `reddit-translator-pro-auto.user.js` 为准。
 
-- 当前对外发布、安装和说明，以 `reddit-translator-pro-auto.user.js` 为准
-- 仓库中的修订文件用于展示演进过程，不代表并行维护多个发布入口
+## 开发与发布约定
+
+1. 新功能、修复和结构调整优先修改 `Reddit 翻译器 Pro 修6.js`
+2. 需要对外发布时，再整理生成 `reddit-translator-pro-auto.user.js`
+3. 发布版本号、README 说明和 `更新日志.md` 需要同步更新
+4. 历史阶段文件不再放回根目录，继续留在 `archive/`
+
+## 本地验证
+
+仓库内已经保留本地验证工具：
+
+- `npm run pw:smoke`
+  运行 Playwright 冒烟测试
+- `npm run browser:chrome`
+  使用脚本启动 Chrome
+- `npm run browser:edge`
+  使用脚本启动 Edge
+
+如果只是安装脚本给浏览器使用，不需要这些 Node.js 工具；这些内容仅用于本地验证和回归检查。
+
+## 历史文件
+
+- `archive/Reddit 翻译器 Pro 修1.js`
+- `archive/Reddit 翻译器 Pro 修2.js`
+- `archive/Reddit 翻译器 Pro 修3.js`
+- `archive/Reddit 翻译器 Pro 修4.js`
+- `archive/Reddit 翻译器 Pro 修5.js`
+- `archive/reddit-translator-v1_0_0.js`
+
+这些文件用于保留脚本演进过程，不作为当前正式安装入口。
+
+## 维护原则
+
+- 对外发布、安装和说明，以 `reddit-translator-pro-auto.user.js` 为准
+- 日常开发和调试，以 `Reddit 翻译器 Pro 修6.js` 为准
+- 历史快照只做归档和对照，不再作为并行维护入口
 - 变更记录请参考 [更新日志.md](./更新日志.md)
