+++ 更新日志.md
@@ -1,125 +1,108 @@
 # Reddit 翻译器 Pro 更新日志
 
-本文档记录仓库内从 `修1` 到 `修5` 的功能演进，并说明当前正式发布文件 `reddit-translator-pro-auto.user.js` 的版本状态。
+本文档记录当前发布线、开发源关系，以及仓库内各阶段版本的主要变化。
 
-## 当前发布状态
+## 当前版本
 
-- 正式发布文件：`reddit-translator-pro-auto.user.js`
-- 当前对应版本：`v1.0.7`
-- 当前代码来源：已与 `Reddit 翻译器 Pro 修5.js` 完全同步
+- 当前开发源：`Reddit 翻译器 Pro 修6.js`
+- 当前正式发布文件：`reddit-translator-pro-auto.user.js`
+- 当前正式发布版本：`v1.0.8`
 
----
+## v1.0.8
 
-## v1.0.7 (修5 / auto.user)
+这是当前正式发布版本。此版本对应当前发布产物 `reddit-translator-pro-auto.user.js`，开发基线为 `Reddit 翻译器 Pro 修6.js`。
 
-本次版本是在 `v1.0.6` 的基础上完成发布整理与持久化隔离修正，当前仓库中的 `reddit-translator-pro-auto.user.js` 与 `Reddit 翻译器 Pro 修5.js` 内容一致。
+### 主要调整
 
-### 主要更新
+- 发布文件版本提升到 `1.0.8`
+- 明确区分“开发源文件”和“正式发布文件”
+- 默认配置针对开箱即用体验重新整理
+- 自动翻译与请求限制的默认参数更偏向直接可用
+- 部分界面文案和快捷键提示被简化为发布导向形式
 
-- 版本号提升到 `1.0.7`
-- `supportURL` 更新为 `reddit-translator-anto/issues`
-- 脚本存储前缀由修4阶段的 `PREFIX = '_x4_'` 调整为 `PREFIX = '_x4c_'`
-- 正式发布文件统一为 `reddit-translator-pro-auto.user.js`
+### 发布线状态
 
-### 影响说明
+- 对外安装入口固定为 `reddit-translator-pro-auto.user.js`
+- `修6` 继续作为当前开发源保留在仓库根目录
+- 历史快照迁移到 `archive/`，不再干扰当前发布入口
 
-- 通过新的存储前缀隔离修4阶段遗留配置，减少旧缓存、旧设置、旧历史记录对当前版本的干扰
-- 便于将修5作为正式发布基线继续维护
+## v1.0.8 / 修6
 
----
+`修6` 是当前开发源，不等同于单独对外发布版本。它承担的是“继续开发和整理发布产物”的角色。
 
-## v1.0.6 (修4)
+### 当前已知定位
 
-这一版是功能层面的关键升级，脚本从“基础翻译工具”进入“可控请求与自动滚动翻译”的阶段。
+- 修复若干持久化键遗漏前缀的问题
+- 作为当前开发主线保留在根目录
+- 用于后续功能迭代、调试和发布整理
 
-### 新增
+## v1.0.5 / 修5
+
+这一阶段主要完成发布整理和持久化隔离，是当前发布线之前的重要基线。
+
+### 主要变化
+
+- 版本提升到 `1.0.5`
+- 整理正式发布入口
+- 调整脚本存储前缀，降低旧配置对新版本的污染
+- 将 `reddit-translator-pro-auto.user.js` 作为正式发布文件固定下来
+
+## v1.0.4 / 修4
+
+修4是功能层面的关键升级，脚本从基础翻译工具进入“请求可控 + 自动滚动翻译”的阶段。
+
+### 主要变化
 
 - 新增请求限制设置
 - 新增长文本自动分块翻译
 - 新增多段文本分批翻译
 - 新增原文 / 译文视图切换
-- 新增自动滚动时的自动翻译触发
-- 新增设置面板中的请求控制项
+- 强化滚动场景中的自动翻译触发逻辑
+- 将请求流程升级为并发与速率可控的调度模型
 
-### 优化
+## v1.0.3 / 修3
 
-- 请求模型从简单串行执行，调整为受并发数和速率限制控制的调度方式
-- 长文本在翻译前先按字符数和段落数切块，降低单次请求失败率
-- 自动翻译加入视口检测、稳定等待和按钮忙碌状态判断，减少重复触发
-- Google、MyMemory、DeepL 统一接入分块翻译流程
-
-### 适合的使用场景
-
-- 长帖正文翻译
-- 多段评论批量翻译
-- 连续滚动浏览时的自动翻译
-- 双语阅读和原文回看
-
----
-
-## v1.0.0 (修3)
-
-修3属于整合阶段，重点不是加入全新引擎，而是把前两版能力收拢为更完整的使用体验。
+修3属于整合阶段，重点是把前两版能力收拢成更稳定的整体结构。
 
 ### 主要变化
 
 - 整合已有翻译能力与面板逻辑
 - 调整默认配置
-- 将自动滚动翻译默认状态从开启改为关闭
-- 继续保留 DeepL Key 配置能力
+- 为修4的大型功能升级准备结构基础
 
-### 阶段意义
+## v1.0.2 / 修2
 
-- 为修4的大型功能升级提供较稳定的结构基础
+修2的重点是补上 `DeepL` 支持，让脚本从双引擎扩展到三引擎方案。
 
----
-
-## v1.0.0 (修2)
-
-修2的重点是补上 DeepL 支持，让脚本从双引擎扩展到三引擎方案。
-
-### 主要更新
+### 主要变化
 
 - 新增 `DeepL` 翻译引擎
 - 新增 `api.deepl.com` 与 `api-free.deepl.com` 连接权限
-- 新增 `deeplApiKeys` 配置项
-- 完善多引擎切换使用场景
+- 新增 `DeepL API Key` 配置能力
 
-### 阶段意义
+## v1.0.1 / 修1
 
-- 从 `Google + MyMemory` 扩展到 `Google + MyMemory + DeepL`
-- 为后续请求控制和分块翻译打下接口基础
-
----
-
-## v1.0.0 (修1)
-
-修1是仓库中的起始版本，已经具备较完整的 Reddit 翻译器形态。
+修1是仓库中的起始版本，已经具备 Reddit 翻译器的基础形态。
 
 ### 初始能力
 
-- 支持 `Google` 与 `MyMemory` 两种翻译引擎
-- 支持 100+ 目标语言
-- 支持主题、双语模式、TTS、快捷键
-- 支持历史记录、转换器与扩展功能入口
-- 支持滚动场景中的元素观察与翻译按钮挂载
-
-### 阶段意义
-
-- 建立脚本的基础架构、界面形态和核心交互方式
-
----
+- 支持 `Google` 与 `MyMemory`
+- 支持多语言目标翻译
+- 支持主题、双语模式、`TTS`、快捷键
+- 支持历史记录和扩展功能入口
 
 ## 文件版本线说明
 
-- `Reddit 翻译器 Pro 修1.js`：初始基础版本
-- `Reddit 翻译器 Pro 修2.js`：新增 DeepL 支持
-- `Reddit 翻译器 Pro 修3.js`：整合阶段版本
-- `Reddit 翻译器 Pro 修4.js`：请求控制、分块翻译、自动滚动翻译强化
-- `Reddit 翻译器 Pro 修5.js`：发布整理与持久化隔离修正版本
-- `reddit-translator-pro-auto.user.js`：当前正式发布文件，已与修5同步
+- `Reddit 翻译器 Pro 修6.js`
+  当前开发源
+- `reddit-translator-pro-auto.user.js`
+  当前正式发布文件
+- `archive/Reddit 翻译器 Pro 修1.js` 到 `archive/Reddit 翻译器 Pro 修5.js`
+  历史阶段快照
+- `archive/reddit-translator-v1_0_0.js`
+  更早期的历史保留文件
 
 ## 备注
 
-- `reddit-translator-v1_0_0.js` 为仓库内保留的早期历史文件，用于对照，不作为当前主发布入口
-- 后续对外安装、分发与说明，均应以 `reddit-translator-pro-auto.user.js` 为准
+- 后续对外安装、分发和说明，均应以 `reddit-translator-pro-auto.user.js` 为准
+- 后续功能开发和调试，均应以 `Reddit 翻译器 Pro 修6.js` 为准
