+++ 浏览器实测矩阵.md
@@ -1,0 +1,297 @@
+# Reddit Translator Pro Auto 浏览器实测矩阵
+
+适用文件：
+
+- `reddit-translator-pro-auto.user.js`
+
+目标：
+
+- 覆盖翻译引擎、内容类型、模式功能和关键交叉路径
+- 验证修复后的配置导入和面板位置持久化
+- 输出每项用例的 `通过 / 失败 / 阻塞` 结果与备注
+
+## 测试前准备
+
+环境要求：
+
+- 浏览器已安装 `Tampermonkey` 或 `Violentmonkey`
+- 已安装最新 `reddit-translator-pro-auto.user.js`
+- 可访问 `https://www.reddit.com/`
+- 如需测 `DeepL`，准备至少 2 个可用 API Key
+- 浏览器允许 `speechSynthesis` 与剪贴板权限
+
+建议测试页面：
+
+- 首页流：用于验证标题、卡片正文、滚动自动翻译
+- 单帖页：用于验证帖子正文、评论、翻译全部、重试
+- 长文帖：正文超过 `maxTextLengthPerRequest`
+- 多段帖：正文或评论包含 6 段以上
+- 中文帖或评论：用于验证语言检测跳过
+
+建议固定测试文本：
+
+- 英文短标题：验证标题翻译
+- 英文长正文：验证分块与长文本
+- 英文多段评论：验证段落分块
+- 中文正文：验证自动滚动 + 跳过
+
+默认记录格式：
+
+- 结果：`通过 / 失败 / 阻塞`
+- 备注：记录页面 URL、引擎、目标语言、异常提示、控制台报错
+
+## 通用基线
+
+每轮正式测试前先执行一次：
+
+1. 打开 Reddit 页面，按 `F2` 确认设置面板可打开。
+2. 目标语言固定为 `zh`。
+3. 先关闭 `隐身模式`、关闭 `自动滚动翻译`。
+4. 清空缓存。
+5. 刷新页面，确认按钮重新注入。
+6. 若测 `DeepL`，先点击 `TEST DEEPL`，确认可用。
+
+## 引擎维度
+
+### E1 Google 基础翻译
+
+- 前置：引擎设为 `Google`
+- 步骤：
+1. 在首页选择 1 个英文标题点击翻译按钮。
+2. 在单帖页选择 1 段英文正文点击翻译按钮。
+3. 在评论区选择 1 条英文评论点击翻译按钮。
+- 预期：
+1. 三类内容都生成译文。
+2. 每个已翻译条目都出现工具栏：复制、TTS、重试。
+3. 历史记录新增对应条目。
+
+### E2 MyMemory 基础翻译
+
+- 前置：引擎设为 `MyMemory`
+- 步骤：重复 `E1`
+- 预期：
+1. 三类内容都可翻译。
+2. 页面无明显卡死。
+3. 历史与统计正常增长。
+
+### E3 DeepL 基础翻译
+
+- 前置：引擎设为 `DeepL`，至少 1 个有效 Key
+- 步骤：重复 `E1`
+- 预期：
+1. 三类内容都可翻译。
+2. 不支持语言时出现明确提示，不应静默失败。
+3. `TEST DEEPL` 显示可用状态。
+
+## 内容维度
+
+### C1 帖子标题
+
+- 步骤：
+1. 在首页找 3 个不同帖子标题逐个翻译。
+- 预期：
+1. 按钮插入位置稳定。
+2. 标题被替换或双语显示符合当前模式。
+3. 切换原文/译文时标题可恢复。
+
+### C2 帖子正文
+
+- 步骤：
+1. 打开单帖页，翻译正文。
+- 预期：
+1. 正文完整翻译。
+2. 工具栏可操作。
+3. 统计与历史增长一次。
+
+### C3 评论
+
+- 步骤：
+1. 翻译至少 3 条不同层级评论。
+- 预期：
+1. 顶层与子评论都能注入按钮。
+2. 翻译后布局不遮挡 Reddit 原有交互。
+
+### C4 长文本
+
+- 前置：将 `maxTextLengthPerRequest` 调低到容易触发分块的值，如 `300`
+- 步骤：
+1. 翻译一段明显超过限制的正文。
+- 预期：
+1. 长文本仍能完整得到译文。
+2. 分块后拼接自然，不缺段、不重复。
+3. 无明显截断。
+
+### C5 多段文本
+
+- 前置：将 `maxParagraphsPerRequest` 调低到 `2`
+- 步骤：
+1. 翻译 4 段以上正文或评论。
+- 预期：
+1. 所有段落都被翻译。
+2. 段间换行结构基本保留。
+
+## 模式维度
+
+### M1 双语开关
+
+- 步骤：
+1. 关闭双语模式，翻译一条正文。
+2. 开启双语模式，刷新页面，再翻译同类正文。
+- 预期：
+1. 关闭时只显示当前视图文本。
+2. 开启时保留原文与译文对照区域。
+
+### M2 原文 / 译文切换
+
+- 步骤：
+1. 翻译至少 2 条内容。
+2. 点击全局视图切换按钮。
+3. 再点击一次切回。
+- 预期：
+1. 所有已翻译条目统一切换。
+2. 不应丢失工具栏和按钮状态。
+
+### M3 自动滚动翻译
+
+- 前置：开启 `Auto-translate on scroll`
+- 步骤：
+1. 打开首页并缓慢下滚。
+2. 观察新进入视口的帖子或评论。
+- 预期：
+1. 元素进入视口后自动开始翻译。
+2. 离开视口前未触发的延时任务会被取消。
+3. 不应对同一元素重复触发多次。
+
+### M4 翻译全部
+
+- 步骤：
+1. 在评论较多的单帖页点击 `TRANSLATE ALL` 或使用热键。
+- 预期：
+1. 当前页所有可翻译条目逐步进入翻译状态。
+2. 进度条或数量变化符合预期。
+3. 已翻译条目不会被重复重新翻译。
+
+### M5 重试
+
+- 步骤：
+1. 翻译任意条目。
+2. 点击该条目的 `Retry`。
+- 预期：
+1. 旧译文被清理。
+2. 重新请求后生成新结果。
+3. 工具栏被正确重建。
+
+### M6 缓存命中
+
+- 步骤：
+1. 翻译同一条内容。
+2. 刷新页面或切回原文后再次翻译同一内容。
+- 预期：
+1. 第二次结果明显更快。
+2. 不出现重复历史污染以外的异常。
+
+### M7 隐身模式
+
+- 步骤：
+1. 开启 `Incognito`
+2. 翻译多条内容。
+3. 打开历史页，刷新页面后再次检查。
+- 预期：
+1. 历史不新增或刷新后不保留。
+2. 缓存不持久化。
+3. 基本翻译能力仍可用。
+
+## 交叉维度
+
+### X1 DeepL + 多 Key 轮换
+
+- 前置：
+1. 配置 2 个以上 DeepL Key
+2. 其中至少 1 个可用
+- 步骤：
+1. 使用 `TEST DEEPL` 确认可用。
+2. 连续翻译多条长内容。
+3. 如果可控，临时让首个 Key 失效，再继续翻译。
+- 预期：
+1. 有效 Key 能继续完成翻译。
+2. 单个 Key 异常时脚本会尝试后续 Key。
+3. 不会因为第一个 Key 失败导致整体直接失效。
+
+### X2 自动滚动翻译 + 语言检测跳过
+
+- 前置：开启 `Auto-translate on scroll`
+- 步骤：
+1. 准备中文内容和英文内容同时出现在页面中。
+2. 滚动让两者都进入视口。
+- 预期：
+1. 英文内容自动翻译。
+2. 中文内容被跳过，不应误翻。
+3. 被跳过的元素不应持续反复触发。
+
+### X3 缓存 + 重试
+
+- 步骤：
+1. 首次翻译一条内容。
+2. 再次翻译确认缓存命中。
+3. 点击 `Retry` 强制重试。
+- 预期：
+1. `Retry` 会清除该条目的缓存再重新请求。
+2. 重试后的结果可以重新进入缓存。
+3. 不会出现“点了重试但仍直接读旧缓存”的情况。
+
+### X4 导入导出 + 热键 + 面板位置
+
+- 步骤：
+1. 打开面板，修改目标语言、引擎、快捷键、颜色。
+2. 拖动面板到新位置并关闭。
+3. 导出配置。
+4. 手动改动设置为另一组值。
+5. 导入刚导出的配置并刷新页面。
+6. 使用导入后的热键打开面板和执行翻译全部。
+- 预期：
+1. 导入后的目标语言、引擎、快捷键、颜色全部恢复。
+2. 面板位置在刷新后保持。
+3. 新热键立即或刷新后按预期生效。
+4. 这一步用于回归本次修复。
+
+### X5 TTS + 已翻译工具栏
+
+- 步骤：
+1. 翻译任意条目。
+2. 点击工具栏的 `Speak`。
+3. 再切换原文/译文视图后再次点击 `Speak`。
+- 预期：
+1. TTS 可开始朗读。
+2. 再次点击或切换状态不会导致工具栏失效。
+3. 同页多个条目反复操作时不应卡死。
+
+## 建议回归顺序
+
+推荐执行顺序：
+
+1. `E1 -> C1/C2/C3 -> M1/M2`
+2. `C4/C5 -> M5/M6`
+3. `E2 -> E3`
+4. `M3/M4/M7`
+5. `X1 -> X5`
+6. 最后执行 `X4`
+
+## 结果记录模板
+
+可按下面格式逐项记录：
+
+```text
+[用例ID] E1
+结果：通过
+环境：Chrome + Tampermonkey
+页面：https://www.reddit.com/...
+备注：标题/正文/评论均成功，历史与统计正常
+```
+
+```text
+[用例ID] X4
+结果：失败
+环境：Edge + Violentmonkey
+页面：https://www.reddit.com/...
+备注：导入后热键恢复正常，但面板位置刷新后丢失
+```
